Add simple GCode Postprocessing plugin
This Will add a comment between each layer.
outputting what the last line of the previous layers was.
Additionally, it will also show if `jerk_enable` was set.

These things will allow us to test if the handshake
happens correctly, if the broadcast occurs and if
the layers can be aware of each other.

Contribute to CURA-10625